refactor: PurgomalumClient common interst move to support package
